Make it compile on scala3

+++ lamp-core/src/main/scala/lamp/nn/MLP.scala
@@ -71,10 +71,11 @@ object MLP {
           EitherModule(
             Right(
               lamp.nn.LayerNorm(
-                normalizedShape = List(normDim),
+                normalizedShape = List(normDim.toLong),
                 tOpt = tOpt,
                 bias = bias,
                 scale = scale
+                
               )
             )
           )
